refactor: extract bundle property restoration
